requirements.txt
tabula-py
pandas
openpyxl
django
